It seems http2 errors are not pointer errors
The type switch used to match to pointers of the http2.*Error types
which given some production data and a test for one, is that neither one
of the three is actually returned as pointer.

Having tests for GoAway and Connection errors will be nice but I have so
far been able to do so.